refactor: Person 분기
