+++ src/components/Card/index.tsx
@@ -3,7 +3,8 @@ import React, { useState } from 'react';
 import { Button } from '../ui/button';
 import Liked from '@/components/Card/Liked';
 import Description from '@/components/Card/Description';
-import ProgressPercentage from '@/components/Card/ProgressPercentage/ProgressPercentage';
+import ProgressPercentage from '@/components/Card/ProgressPercentage';
+import Person from '@/components/Card/Person';
 
 export default function Card({ data }: any) {
   const [isBookmarked, setIsBookmarked] = useState(false);
@@ -11,8 +12,6 @@ export default function Card({ data }: any) {
   const handleClick = () => {
     setIsBookmarked((prevIsBookmarked) => !prevIsBookmarked);
   };
-
-  // const progressPercentage = (data.member / 20) * 100;
 
   return (
     <div className="h-230 relative flex w-full gap-20 rounded-lg bg-white p-20">
@@ -24,33 +23,24 @@ export default function Card({ data }: any) {
           </div>
         )}
       </div>
+
       <div className="gap-30 relative flex w-full flex-col items-start justify-between text-gray-600">
         <Description data={data} />
 
         <div className="mb-11 flex w-full items-center justify-center gap-16">
           {data.member > 5 ? (
             <>
-              <Image src="/icons/overPersonIcon.svg" alt="overPersonIcon" width={24} height={24} />
-              <div className="text-14 text-primary-300">{data.member}/20</div>
+              <Person data={data} />
+              {/* <Image src="/icons/overPersonIcon.svg" alt="overPersonIcon" width={24} height={24} />
+              <div className="text-14 text-primary-300">{data.member}/20</div> */}
               <ProgressPercentage data={data} />
-              {/* <div className="relative h-6 w-full rounded-full bg-gray-100">
-                <div
-                  className="bg-primary-300 absolute h-6 rounded-full"
-                  style={{ width: `${progressPercentage}%` }}
-                />
-              </div> */}
             </>
           ) : (
             <>
-              <Image src="/icons/basePersonIcon.svg" alt="basePersonIcon" width={24} height={24} />
-              <div className="text-14 text-neutral-500">{data.member}/20</div>
+              <Person data={data} />
+              {/* <Image src="/icons/basePersonIcon.svg" alt="basePersonIcon" width={24} height={24} />
+              <div className="text-14 text-neutral-500">{data.member}/20</div> */}
               <ProgressPercentage data={data} />
-              {/* <div className="relative h-6 w-full rounded-full bg-gray-100">
-                <div
-                  className="bg-primary-300 absolute h-6 rounded-full"
-                  style={{ width: `${progressPercentage}%` }}
-                />
-              </div> */}
             </>
           )}
         </div>
